regorganize main.py split to model_utils.py and run_on_example_prompts.py
